Move connectors to dbml-core
