+++ packages/dbml-cli/package.json
@@ -33,9 +33,7 @@
     "esm": "^3.2.25",
     "figures": "^3.2.0",
     "lodash": "^4.17.15",
-    "mssql": "^11.0.1",
     "pegjs-require-import": "^0.0.2",
-    "pg": "^8.12.0",
     "strip-ansi": "^5.2.0",
     "winston": "^3.2.1"
   },
