+++ packages/dbml-cli/src/cli/connector.js
@@ -2,7 +2,6 @@ import { importer } from '@dbml/core';
 import figures from 'figures';
 import chalk from 'chalk';
 import path from 'path';
-import { fetchSchemaJson } from '../connectors/Connector';
 import {
   resolvePaths,
   getConnectionOpt,
@@ -16,13 +15,12 @@ export default async function connectionHandler (program) {
     const connection = program.args[0];
     const opts = program.opts();
     const { format } = getConnectionOpt(opts);
-    const schemaJson = await fetchSchemaJson(connection, format);
 
     if (!opts.outFile && !opts.outDir) {
-      const res = importer.generateDbml(schemaJson);
+      const res = await importer.generateDbml(connection, format);
       OutputConsolePlugin.write(res);
     } else if (opts.outFile) {
-      const res = importer.generateDbml(schemaJson);
+      const res = await importer.generateDbml(connection, format);
       (new OutputFilePlugin(resolvePaths(opts.outFile))).write(res);
 
       // bearer:disable javascript_lang_logger
