+++ packages/dbml-core/package.json
@@ -35,6 +35,7 @@
     "@dbml/parse": "^3.6.0",
     "antlr4": "^4.13.1",
     "lodash": "^4.17.15",
+    "mssql": "^11.0.1",
     "parsimmon": "^1.13.0",
     "pg": "^8.12.0",
     "pluralize": "^8.0.0"
