+++ packages/dbml-core/src/import/index.js
@@ -1,4 +1,5 @@
-import { generateDatabase } from '../parse/DatabaseGenerator';
+import { generateDatabase } from '../parse/databaseGenerator';
+import { fetchSchemaJson } from '../connectors/connector';
 import Parser from '../parse/Parser';
 import ModelExporter from '../export/ModelExporter';
 
@@ -9,7 +10,8 @@ function _import (str, format) {
   return dbml;
 }
 
-function generateDbml (schemaJson) {
+async function generateDbml (connection, format) {
+  const schemaJson = await fetchSchemaJson(connection, format);
   const database = generateDatabase(schemaJson);
   const dbml = ModelExporter.export(database.normalize(), 'dbml');
