+++ packages/dbml-core/types/import/index.d.ts
@@ -1,5 +1,5 @@
 declare function _import(str: string, format: 'dbml' | 'mysql' | 'postgres' | 'json' | 'mssql' | 'postgresLegacy'): string;
-declare function generateDbml(schemaJson: any): string;
+declare async function generateDbml(connection: any, format: 'mysql' | 'postgres' | 'mssql'): string;
 declare const _default: {
     import: typeof _import;
     generateDbml: typeof generateDbml;
